backend/requirements.txt
# Unbot AI Backend - Python 依赖

# Web 框架
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 数据验证
pydantic==2.6.1
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.25
aiomysql==0.2.0

# Redis
redis==5.0.1

# HTTP 客户端
httpx==0.26.0

# JWT 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# 环境变量
python-dotenv==1.0.1

# 工具库
python-dateutil==2.8.2

# 格式化
black==24.1.1
isort==5.13.2
